Feito o layout principal da Home Page, ainda hardcoded com placeholders, sem puxar informações do back / bd. Carrossel principal de pets e posts na home.
